Massive overhaul of headers: (part 1) - Using forward declaration where possible - Getting rid of includes - Logger class moved to it's own source/header files - "packing includes headers" GameSystem.h, GraphicsSystem.h and SDLStuff.h removed from project; files should forward declare/include only when needed - linux.h removed; platforms will be handled in better way - Various code tidying; Moving * and & to type names - Moved some class methods from headers to source files

+++ trunk/WhoreMasterRenewal/src/libRocketSFMLInterface/SystemInterfaceSFML.cpp
@@ -24,9 +24,12 @@
  * THE SOFTWARE.
  *
  */
-#include <iostream>
-#include <Rocket/Core.h>
+
 #include "SystemInterfaceSFML.h"
+
+#include <Rocket/Core.h>
+
+#include <iostream>
 
 int RocketSFMLSystemInterface::GetKeyModifiers( sf::Event& event )
 {
@@ -381,7 +384,7 @@ Rocket::Core::Input::KeyIdentifier RocketSFMLSystemInterface::TranslateKey( sf::
         break;
     // report default case as Unknown key too
     default:
-        std::cout << "default: keycode=" << key << "\n";
+        std::cout << "default: keycode=" << static_cast<long int>(key) << "\n";
         return Rocket::Core::Input::KeyIdentifier::KI_UNKNOWN;
         break;
 	};
